fix: Shared Word Card uses full width
